requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
pypdf
ollama
streamlit
